applix.c: Fixed regression (nw)
git-svn-id: svn://messdev.no-ip.org/mess@14919 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/drivers/applix.c
@@ -40,7 +40,7 @@ public:
 	DECLARE_READ16_MEMBER(applix_inputs_r);
 	DECLARE_WRITE16_MEMBER(applix_index_w);
 	DECLARE_WRITE16_MEMBER(applix_register_w);
-	required_shared_ptr<UINT8> m_base;
+	required_shared_ptr<UINT16> m_base;
 };
 
 WRITE16_MEMBER( applix_state::applix_index_w )
